+++ dependencies.gradle
@@ -22,6 +22,6 @@
  * For more details, see https://docs.gradle.org/7.6/userguide/java_library_plugin.html#sec:java_library_configurations_graph
  */
 dependencies {
-    api("com.github.GTNewHorizons:GT5-Unofficial:5.09.49.03:dev")
+    api("com.github.GTNewHorizons:GT5-Unofficial:5.09.49.55:dev")
     api("com.github.GTNewHorizons:CraftTweaker:3.3.1:dev")
 }
