+++ src/main/java/gttweaker/mods/gregtech/Fuels.java
@@ -1,10 +1,10 @@
 package gttweaker.mods.gregtech;
 
-import static gregtech.api.util.GT_RecipeConstants.FUEL_TYPE;
-import static gregtech.api.util.GT_RecipeConstants.FUEL_VALUE;
+import static gregtech.api.util.GTRecipeConstants.FUEL_TYPE;
+import static gregtech.api.util.GTRecipeConstants.FUEL_VALUE;
 
-import gregtech.api.enums.GT_Values;
-import gregtech.api.util.GT_RecipeConstants;
+import gregtech.api.enums.GTValues;
+import gregtech.api.util.GTRecipeConstants;
 import gttweaker.mods.AddMultipleRecipeAction;
 import minetweaker.MineTweakerAPI;
 import minetweaker.annotations.ModOnly;
@@ -115,14 +115,14 @@ public class Fuels {
 
         @Override
         protected void applySingleRecipe(ArgIterator i) {
-            GT_Values.RA.stdBuilder()
+            GTValues.RA.stdBuilder()
                 .itemInputs(i.nextItem())
                 .itemOutputs(i.nextItem())
                 .metadata(FUEL_VALUE, i.nextInt())
                 .metadata(FUEL_TYPE, i.nextInt())
                 .duration(0)
                 .eut(0)
-                .addTo(GT_RecipeConstants.Fuel);
+                .addTo(GTRecipeConstants.Fuel);
         }
     }
 }
