+++ src/main/java/gttweaker/mods/gregtech/machines/Autoclave.java
@@ -1,8 +1,8 @@
 package gttweaker.mods.gregtech.machines;
 
-import static gregtech.api.enums.GT_Values.RA;
+import static gregtech.api.enums.GTValues.RA;
 import static gregtech.api.recipe.RecipeMaps.autoclaveRecipes;
-import static gregtech.api.util.GT_RecipeConstants.LOW_GRAVITY;
+import static gregtech.api.util.GTRecipeConstants.LOW_GRAVITY;
 
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.fluids.FluidStack;
