+++ src/main/java/gttweaker/mods/gregtech/machines/DistillationTower.java
@@ -1,13 +1,13 @@
 package gttweaker.mods.gregtech.machines;
 
-import static gregtech.api.enums.GT_Values.RA;
+import static gregtech.api.enums.GTValues.RA;
 import static gregtech.api.recipe.RecipeMaps.distillationTowerRecipes;
 import static gregtech.api.recipe.RecipeMaps.distilleryRecipes;
 
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.fluids.FluidStack;
 
-import gregtech.api.util.GT_Utility;
+import gregtech.api.util.GTUtility;
 import gttweaker.mods.AddMultipleRecipeAction;
 import minetweaker.MineTweakerAPI;
 import minetweaker.annotations.ModOnly;
@@ -92,7 +92,7 @@ public class DistillationTower {
 
                     for (int idx = 0; idx < Math.min(fluidOutputs.length, 11); idx++) {
                         RA.stdBuilder()
-                            .itemInputs(GT_Utility.getIntegratedCircuit(idx + 1))
+                            .itemInputs(GTUtility.getIntegratedCircuit(idx + 1))
                             .itemOutputs(output)
                             .fluidInputs(fluidInput)
                             .fluidOutputs(fluidOutputs[idx])
