+++ src/main/java/gttweaker/mods/gregtech/machines/OilCracker.java
@@ -1,9 +1,9 @@
 package gttweaker.mods.gregtech.machines;
 
-import static gregtech.api.enums.GT_Values.RA;
+import static gregtech.api.enums.GTValues.RA;
 import static gregtech.api.recipe.RecipeMaps.crackingRecipes;
 
-import gregtech.api.util.GT_Utility;
+import gregtech.api.util.GTUtility;
 import gttweaker.mods.AddMultipleRecipeAction;
 import minetweaker.MineTweakerAPI;
 import minetweaker.annotations.ModOnly;
@@ -45,7 +45,7 @@ public class OilCracker {
                 @Override
                 protected void applySingleRecipe(ArgIterator i) {
                     RA.stdBuilder()
-                        .itemInputs(GT_Utility.getIntegratedCircuit(i.nextInt()))
+                        .itemInputs(GTUtility.getIntegratedCircuit(i.nextInt()))
                         .fluidInputs(i.nextFluid(), i.nextFluid())
                         .fluidOutputs(i.nextFluid())
                         .duration(i.nextInt())
