+++ src/main/java/gttweaker/mods/gtpp/machines/BlastSmelter.java
@@ -1,6 +1,6 @@
 package gttweaker.mods.gtpp.machines;
 
-import static gregtech.api.enums.GT_Values.RA;
+import static gregtech.api.enums.GTValues.RA;
 import static gtPlusPlus.api.recipe.GTPPRecipeMaps.alloyBlastSmelterRecipes;
 
 import net.minecraft.item.ItemStack;
