+++ src/main/java/gttweaker/mods/gtpp/machines/MultiblockElectrolyzer.java
@@ -1,6 +1,6 @@
 package gttweaker.mods.gtpp.machines;
 
-import static gregtech.api.enums.GT_Values.RA;
+import static gregtech.api.enums.GTValues.RA;
 import static gregtech.api.recipe.RecipeMaps.electrolyzerNonCellRecipes;
 
 import gttweaker.mods.AddMultipleRecipeAction;
